Moved out some middleware and plugins into custom section. These might later become an extensios nuget package
