+++ src/Noobot.Core/DependencyResolution/ContainerFactory.cs
@@ -98,16 +98,16 @@ namespace Noobot.Core.DependencyResolution
                 Type nextType = pipeline.Pop();
                 var nextDeclare = registry.For<IMiddleware>();
 
+                // defined here so they can be overridden
+                registry.For<IMiddleware>().DecorateAllWith<HelpMiddleware>();
+                registry.For<IMiddleware>().DecorateAllWith<AboutMiddleware>();
+                registry.For<IMiddleware>().DecorateAllWith<StatsMiddleware>();
+
                 MethodInfo decorateMethod = nextDeclare.GetType().GetMethod("DecorateAllWith", new[] { typeof(Func<Instance, bool>) });
                 MethodInfo generic = decorateMethod.MakeGenericMethod(nextType);
                 generic.Invoke(nextDeclare, new object[] { null });
             }
 
-            registry.For<IMiddleware>().DecorateAllWith<AboutMiddleware>();
-            registry.For<IMiddleware>().DecorateAllWith<ScheduleMiddleware>();
-            registry.For<IMiddleware>().DecorateAllWith<StatsMiddleware>();
-            registry.For<IMiddleware>().DecorateAllWith<AdminMiddleware>();
-            registry.For<IMiddleware>().DecorateAllWith<HelpMiddleware>();
             registry.For<IMiddleware>().DecorateAllWith<BeginMessageMiddleware>();
         }
 
@@ -128,10 +128,7 @@ namespace Noobot.Core.DependencyResolution
         {
             var pluginTypes = new List<Type>
             {
-                typeof (StoragePlugin),
-                typeof (SchedulePlugin),
-                typeof (StatsPlugin),
-                typeof (AdminPlugin)
+                typeof (StatsPlugin)
             };
 
             pluginTypes.AddRange(_pluginConfiguration.ListPluginTypes());
