+++ src/Noobot.Core/MessagingPipeline/Middleware/HandlerMapping.cs
@@ -29,13 +29,18 @@ namespace Noobot.Core.MessagingPipeline.Middleware
         public Func<IncomingMessage, string, IEnumerable<ResponseMessage>> EvaluatorFunc { get; set; }
 
         /// <summary>
-        /// Defaults to "False". If set to True then the pipeline isn't interupted if a match occurs here.
+        /// Defaults to "False". If set to True then the pipeline isn't interupted if a match occurs here. This is good for logging.
         /// </summary>
         public bool ShouldContinueProcessing { get; set; }
 
         /// <summary>
-        /// Defaults to "True". If set to false then any message is considered, even if it isn't targeted at the bot.
+        /// Defaults to "True". If set to false then any message is considered, even if it isn't targeted at the bot. e.g. @noobot or a private channel
         /// </summary>
         public bool MessageShouldTargetBot { get; set; }
+
+        /// <summary>
+        /// Defaults to "True". Set to false to hide these commands in the help command.
+        /// </summary>
+        public bool VisibleInHelp { get; set; }
     }
 }
