+++ src/Noobot.Custom/Pipeline/Middleware/AdminMiddleware.cs
@@ -1,11 +1,13 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using Noobot.Core;
 using Noobot.Core.Logging;
+using Noobot.Core.MessagingPipeline.Middleware;
 using Noobot.Core.MessagingPipeline.Request;
 using Noobot.Core.MessagingPipeline.Response;
-using Noobot.Core.Plugins.StandardPlugins;
+using Noobot.Custom.Plugins;
 
-namespace Noobot.Core.MessagingPipeline.Middleware.StandardMiddleware
+namespace Noobot.Custom.Pipeline.Middleware
 {
     internal class AdminMiddleware : MiddlewareBase
     {
