+++ src/Noobot.Tests.Unit/Noobot.Tests.Unit.csproj
@@ -121,6 +121,10 @@
       <Project>{20c88122-7ba3-4380-89fc-350f9d5b94dc}</Project>
       <Name>Noobot.Core</Name>
     </ProjectReference>
+    <ProjectReference Include="..\Noobot.Custom\Noobot.Custom.csproj">
+      <Project>{5794b29f-234c-4b12-aae4-b31363f78c8d}</Project>
+      <Name>Noobot.Custom</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Noobot.Runner\Noobot.Runner.csproj">
       <Project>{fc387cb5-8b7f-4873-8a14-8337682d2c73}</Project>
       <Name>Noobot.Runner</Name>
